refactor(profaastinate): nexus
Make a depcrecated savety copy of the nexus
