+++ profaastinate/deprecated/nexus/bulk/models/bulk-structs.go
