+++ profaastinate/deprecated/nexus/common/models/structs/nexus-item.struct.go
