+++ profaastinate/deprecated/nexus/go.mod
@@ -1,0 +1,3 @@
+module github.com/konsumgandalf/profaastinate/nexus
+
+go 1.21.4
