--- profaastinate/nexus/go.mod
@@ -1,3 +1,0 @@
-module nexus
-
-go 1.21.4
